data_scarpe/requirements.txt
async-generator==1.10
attrs==22.1.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.2
google-images-download @ git+https://github.com/Joeclinton1/google-images-download.git@03671f3f37c4af113af63f650937412d070b3806
h11==0.13.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytube==12.1.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wincertstore==0.2
wsproto==1.2.0
